+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+Unreleased
+------
+* Add support for forecast prices (PRS)
+
 1.1.1 / 2017-05-011
 ------
 * Change URL from www2.electricityinfo.co.nz to electricityinfo.co.nz